refactor: UserService 추가 및 LoginController, RegisterController 개선
