+++ tomcat/src/main/java/nextstep/jwp/presentation/LoginController.java
@@ -4,11 +4,9 @@ import static nextstep.jwp.presentation.StaticResource.INDEX_PAGE;
 import static nextstep.jwp.presentation.StaticResource.UNAUTHORIZED_PAGE;
 import static org.apache.catalina.Session.JSESSIONID;
 
-import java.util.Objects;
-import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.UnauthorizedException;
-import nextstep.jwp.exception.UserNotFoundException;
 import nextstep.jwp.model.User;
+import nextstep.jwp.service.UserService;
 import org.apache.catalina.Session;
 import org.apache.coyote.HttpRequest;
 import org.apache.coyote.HttpResponse;
@@ -18,14 +16,14 @@ import org.slf4j.LoggerFactory;
 
 public class LoginController extends AbstractController {
 
-    private static final LoginController INSTANCE = new LoginController();
+    private static final LoginController INSTANCE = new LoginController(UserService.getInstance());
 
     private static final Logger LOG = LoggerFactory.getLogger(LoginController.class);
 
-    private static final String ACCOUNT_PARAM = "account";
-    private static final String PASSWORD_PARAM = "password";
+    private final UserService userService;
 
-    private LoginController() {
+    private LoginController(final UserService userService) {
+        this.userService = userService;
     }
 
     public static LoginController getInstance() {
@@ -46,9 +44,7 @@ public class LoginController extends AbstractController {
     @Override
     protected void doPost(final HttpRequest request, final HttpResponse response) throws Exception {
         try {
-            final User user = login(request);
-
-            logSuccessMessage(user);
+            final User user = userService.login(request);
             setLoginSession(request, response, user);
             redirectIndex(response);
         } catch (final UnauthorizedException unauthorizedException) {
@@ -65,26 +61,6 @@ public class LoginController extends AbstractController {
             return null;
         }
         return (User) session.getAttribute("user");
-    }
-
-    private User login(final HttpRequest request) {
-        final String account = request.getBodyParam(ACCOUNT_PARAM);
-        final String password = request.getBodyParam(PASSWORD_PARAM);
-
-        Objects.requireNonNull(account, "account는 null이면 안됩니다.");
-        Objects.requireNonNull(password, "password는 null이면 안됩니다.");
-
-        final User user = InMemoryUserRepository.findByAccount(account)
-                .orElseThrow(UserNotFoundException::new);
-        if (!user.checkPassword(password)) {
-            throw new UnauthorizedException();
-        }
-        return user;
-    }
-
-    private void logSuccessMessage(final User user) {
-        final String msg = user.toString();
-        LOG.info(msg);
     }
 
     private void setLoginSession(final HttpRequest request, final HttpResponse response, final User user) {
