+++ tomcat/src/main/java/nextstep/jwp/presentation/RegisterController.java
@@ -3,10 +3,8 @@ package nextstep.jwp.presentation;
 import static nextstep.jwp.presentation.StaticResource.INDEX_PAGE;
 import static org.apache.catalina.Session.JSESSIONID;
 
-import java.util.Objects;
-import java.util.Optional;
-import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
+import nextstep.jwp.service.UserService;
 import org.apache.catalina.Session;
 import org.apache.coyote.HttpRequest;
 import org.apache.coyote.HttpResponse;
@@ -16,15 +14,14 @@ import org.slf4j.LoggerFactory;
 
 public class RegisterController extends AbstractController {
 
-    private static final RegisterController INSTANCE = new RegisterController();
+    private static final RegisterController INSTANCE = new RegisterController(UserService.getInstance());
 
     private static final Logger LOG = LoggerFactory.getLogger(RegisterController.class);
 
-    private static final String ACCOUNT_PARAM = "account";
-    private static final String EMAIL_PARAM = "email";
-    private static final String PASSWORD_PARAM = "password";
+    private final UserService userService;
 
-    private RegisterController() {
+    private RegisterController(final UserService userService) {
+        this.userService = userService;
     }
 
     public static RegisterController getInstance() {
@@ -39,9 +36,7 @@ public class RegisterController extends AbstractController {
     @Override
     protected void doPost(final HttpRequest request, final HttpResponse response) throws Exception {
         try {
-            final User user = register(request);
-
-            logSuccessMessage(user);
+            final User user = userService.register(request);
             setLoginSession(request, response, user);
             redirectIndex(response);
         } catch (final RuntimeException runtimeException) {
@@ -49,37 +44,6 @@ public class RegisterController extends AbstractController {
         }
 
         response.setBody(StaticResource.ofRequest(request));
-    }
-
-    private User register(final HttpRequest request) {
-        final String account = request.getBodyParam(ACCOUNT_PARAM);
-        final String email = request.getBodyParam(EMAIL_PARAM);
-        final String password = request.getBodyParam(PASSWORD_PARAM);
-
-        checkNull(account, email, password);
-        checkAccountDuplication(account);
-
-        final User user = new User(account, password, email);
-        InMemoryUserRepository.save(user);
-        return user;
-    }
-
-    private void checkNull(final String account, final String email, final String password) {
-        Objects.requireNonNull(account, "account는 null이면 안됩니다.");
-        Objects.requireNonNull(email, "email은 null이면 안됩니다.");
-        Objects.requireNonNull(password, "password는 null이면 안됩니다.");
-    }
-
-    private void checkAccountDuplication(final String account) {
-        final Optional<User> savedAccount = InMemoryUserRepository.findByAccount(account);
-        if (savedAccount.isPresent()) {
-            throw new IllegalStateException("중복된 아이디입니다.");
-        }
-    }
-
-    private void logSuccessMessage(final User user) {
-        final String msg = user.toString();
-        LOG.info(msg);
     }
 
     private void setLoginSession(final HttpRequest request, final HttpResponse response, final User user) {
@@ -92,5 +56,4 @@ public class RegisterController extends AbstractController {
         response.setStatus(HttpStatus.FOUND);
         response.setLocation(INDEX_PAGE);
     }
-
 }
